moved call to store for countItems to header compnonent where it is used
